Create link class
